# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Stargate API
  description: |
    Stargate is a high-performance API gateway that provides comprehensive route management, 
    load balancing, plugin system, and developer portal functionality.
    
    ## Authentication
    
    Stargate supports multiple authentication methods:
    - **API Key**: Use `X-Admin-Key` header for admin operations
    - **JWT Bearer**: Use `Authorization: Bearer <token>` for user operations
    
    ## Rate Limiting
    
    API requests are subject to rate limiting:
    - Admin API: 1000 requests/hour per API key
    - Portal API: 100 requests/minute per user
    
    ## Error Handling
    
    All endpoints return consistent error responses with HTTP status codes and JSON error objects.
  version: 1.0.0
  contact:
    name: Stargate Team
    url: https://github.com/stargate/stargate
    email: support@stargate.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:9090
    description: Development server (Admin API)
  - url: http://localhost:8080
    description: Development server (Portal API)
  - url: https://api.stargate.io
    description: Production server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      summary: Authenticate user
      description: Authenticate user credentials and return JWT token
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              admin:
                summary: Admin login
                value:
                  username: admin
                  password: password
              developer:
                summary: Developer login
                value:
                  username: developer
                  password: dev_password
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/api-keys:
    post:
      summary: Generate API key
      description: Generate a new API key for authentication
      operationId: generateApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyRequest'
      responses:
        '200':
          description: API key generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/routes:
    get:
      summary: List routes
      description: Get a paginated list of all routes
      operationId: listRoutes
      parameters:
        - name: limit
          in: query
          description: Maximum number of routes to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          description: Number of routes to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: search
          in: query
          description: Search routes by name or host
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of routes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteListResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create route
      description: Create a new route configuration
      operationId: createRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteRequest'
            examples:
              basic:
                summary: Basic route
                value:
                  name: API Route
                  rules:
                    hosts:
                      - api.example.com
                    paths:
                      - type: prefix
                        value: /api
                    methods:
                      - GET
                      - POST
                  upstream_id: upstream-001
              advanced:
                summary: Advanced route with regex
                value:
                  name: User API Route
                  rules:
                    hosts:
                      - api.example.com
                    paths:
                      - type: regex
                        value: ^/api/v1/users/[0-9]+$
                    methods:
                      - GET
                      - PUT
                      - DELETE
                  upstream_id: upstream-001
                  priority: 100
                  metadata:
                    team: backend
                    environment: production
      responses:
        '201':
          description: Route created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict - route already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/routes/{routeId}:
    parameters:
      - name: routeId
        in: path
        required: true
        description: Unique identifier of the route
        schema:
          type: string
          example: route-001

    get:
      summary: Get route
      description: Get a specific route by ID
      operationId: getRoute
      responses:
        '200':
          description: Route details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update route
      description: Update an existing route
      operationId: updateRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteRequest'
      responses:
        '200':
          description: Route updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete route
      description: Delete a route by ID
      operationId: deleteRoute
      responses:
        '200':
          description: Route deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/upstreams:
    get:
      summary: List upstreams
      description: Get a paginated list of all upstreams
      operationId: listUpstreams
      parameters:
        - name: limit
          in: query
          description: Maximum number of upstreams to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          description: Number of upstreams to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of upstreams
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpstreamListResponse'

    post:
      summary: Create upstream
      description: Create a new upstream configuration
      operationId: createUpstream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpstreamRequest'
            examples:
              basic:
                summary: Basic upstream
                value:
                  name: Backend Service
                  targets:
                    - url: http://backend1:8080
                      weight: 100
                    - url: http://backend2:8080
                      weight: 100
                  algorithm: round_robin
              with_health_check:
                summary: Upstream with health check
                value:
                  name: Backend Service
                  targets:
                    - url: http://backend1:8080
                      weight: 100
                  algorithm: round_robin
                  health_check:
                    enabled: true
                    path: /health
                    interval: 30
                    timeout: 5
      responses:
        '201':
          description: Upstream created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpstreamResponse'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Admin API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
          example: password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: string
          description: Token expiration duration
          example: 24h
        refresh_token:
          type: string
          description: Refresh token for obtaining new access tokens
          example: rt_1234567890abcdef
    ApiKeyRequest:
      type: object
      properties:
        name:
          type: string
          description: Human-readable name for the API key
          example: my-api-key
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully
    Error:
      type: object
      required:
        - error
        - status
      properties:
        error:
          type: string
          description: Error message
          example: Validation failed
        status:
          type: integer
          description: HTTP status code
          example: 400
        details:
          type: string
          description: Additional error details
          example: Field 'name' is required
        code:
          type: string
          description: Error code for programmatic handling
          example: VALIDATION_ERROR
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp of the error
          example: 1640995200
...